[MODORDERS-380] - Can not receive when order is closed
