+++ src/test/java/org/folio/orders/events/handlers/CheckInOrderStatusChangeChangeHandlerTest.java
@@ -19,7 +19,6 @@ import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.Matchers.nullValue;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeoutException;
@@ -29,7 +28,6 @@ import java.util.stream.Stream;
 import javax.ws.rs.core.Response;
 
 import org.folio.orders.utils.HelperUtils;
-import org.folio.rest.impl.AbstractHelper;
 import org.folio.rest.impl.ApiTestBase;
 import org.folio.rest.impl.CheckinHelper;
 import org.folio.rest.jaxrs.model.PurchaseOrder;
@@ -71,7 +69,7 @@ public class CheckInOrderStatusChangeChangeHandlerTest extends ApiTestBase {
   @Test
   public void testUpdateOpenOrderToClosed(TestContext context) {
     logger.info("=== Test case when order status update is expected from Open to Closed ===");
-    sendEvent(createBody(PO_ID_OPEN_TO_BE_CLOSED), context.asyncAssertSuccess(result -> {
+    sendEvent(createBody(false, PO_ID_OPEN_TO_BE_CLOSED), context.asyncAssertSuccess(result -> {
       assertThat(getPurchaseOrderRetrievals(), hasSize(1));
       assertThat(getPoLineSearches(), hasSize(1));
       assertThat(getPurchaseOrderUpdates(), hasSize(1));
@@ -95,7 +93,7 @@ public class CheckInOrderStatusChangeChangeHandlerTest extends ApiTestBase {
   @Test
   public void testUpdateNotRequiredForOpenOrder(TestContext context) {
     logger.info("=== Test case when no order update is expected for Open order ===");
-    sendEvent(createBody(PO_ID_OPEN_STATUS), context.asyncAssertSuccess(result -> {
+    sendEvent(createBody(false, PO_ID_OPEN_STATUS), context.asyncAssertSuccess(result -> {
       assertThat(getPurchaseOrderRetrievals(), hasSize(1));
       assertThat(getPoLineSearches(), hasSize(1));
       assertThat(getPurchaseOrderUpdates(), nullValue());
@@ -106,7 +104,7 @@ public class CheckInOrderStatusChangeChangeHandlerTest extends ApiTestBase {
   @Test
   public void testUpdateClosedOrderToOpen(TestContext context) {
     logger.info("=== Test case when order update is expected for Closed order ===");
-    sendEvent(createBody(PO_ID_CLOSED_STATUS), context.asyncAssertSuccess(result -> {
+    sendEvent(createBody(false, PO_ID_CLOSED_STATUS), context.asyncAssertSuccess(result -> {
       assertThat(getPurchaseOrderRetrievals(), hasSize(1));
       assertThat(getPoLineSearches(), hasSize(1));
       assertThat(getPurchaseOrderUpdates(), hasSize(1));
@@ -126,7 +124,7 @@ public class CheckInOrderStatusChangeChangeHandlerTest extends ApiTestBase {
   @Test
   public void testNoUpdatesForPendingOrderWithoutLines(TestContext context) {
     logger.info("=== Test case when no order update is expected for Pending order without lines ===");
-    sendEvent(createBody(PO_ID_PENDING_STATUS_WITHOUT_PO_LINES), context.asyncAssertSuccess(result -> {
+    sendEvent(createBody(false, PO_ID_PENDING_STATUS_WITHOUT_PO_LINES), context.asyncAssertSuccess(result -> {
       assertThat(getPurchaseOrderRetrievals(), hasSize(1));
       assertThat(getPoLineSearches(), nullValue());
       assertThat(getPurchaseOrderUpdates(), nullValue());
@@ -137,7 +135,7 @@ public class CheckInOrderStatusChangeChangeHandlerTest extends ApiTestBase {
   @Test
   public void testNoUpdatesForPendingOrderWithLines(TestContext context) {
     logger.info("=== Test case when no order update is expected for Pending order with a few PO Lines===");
-    sendEvent(createBody(PO_ID_PENDING_STATUS_WITH_PO_LINES), context.asyncAssertSuccess(result -> {
+    sendEvent(createBody(false, PO_ID_PENDING_STATUS_WITH_PO_LINES), context.asyncAssertSuccess(result -> {
       assertThat(getPurchaseOrderRetrievals(), hasSize(1));
       assertThat(getPoLineSearches(), nullValue());
       assertThat(getPurchaseOrderUpdates(), nullValue());
@@ -148,7 +146,7 @@ public class CheckInOrderStatusChangeChangeHandlerTest extends ApiTestBase {
   @Test
   public void testUpdateClosedOrderToOpenAndNoUpdateForOpenOrder(TestContext context) {
     logger.info("=== Test case when order update is expected for Closed order ===");
-    sendEvent(createBody(PO_ID_CLOSED_STATUS, PO_ID_OPEN_STATUS), context.asyncAssertSuccess(result -> {
+    sendEvent(createBody(false, PO_ID_CLOSED_STATUS, PO_ID_OPEN_STATUS), context.asyncAssertSuccess(result -> {
       assertThat(getPurchaseOrderRetrievals(), hasSize(2));
       assertThat(getPoLineSearches(), hasSize(2));
       assertThat(getPurchaseOrderUpdates(), hasSize(1));
@@ -160,7 +158,7 @@ public class CheckInOrderStatusChangeChangeHandlerTest extends ApiTestBase {
   @Test
   public void testNonexistentOrder(TestContext context) {
     logger.info("=== Test case when no order update is expected ===");
-    sendEvent(createBody(ID_DOES_NOT_EXIST), context.asyncAssertFailure(result -> {
+    sendEvent(createBody(false, ID_DOES_NOT_EXIST), context.asyncAssertFailure(result -> {
       assertThat(getPurchaseOrderRetrievals(), nullValue());
       assertThat(getPoLineSearches(), nullValue());
       assertThat(getPurchaseOrderUpdates(), nullValue());
@@ -172,7 +170,7 @@ public class CheckInOrderStatusChangeChangeHandlerTest extends ApiTestBase {
   @Test
   public void testOrderRetrievalFailure(TestContext context) {
     logger.info("=== Test case when order retrieval fails ===");
-    sendEvent(createBody(ID_FOR_INTERNAL_SERVER_ERROR), context.asyncAssertFailure(result -> {
+    sendEvent(createBody(false, ID_FOR_INTERNAL_SERVER_ERROR), context.asyncAssertFailure(result -> {
       assertThat(getPurchaseOrderRetrievals(), nullValue());
       assertThat(getPoLineSearches(), nullValue());
       assertThat(getPurchaseOrderUpdates(), nullValue());
@@ -184,7 +182,7 @@ public class CheckInOrderStatusChangeChangeHandlerTest extends ApiTestBase {
   @Test
   public void testOrderLinesRetrievalFailure(TestContext context) {
     logger.info("=== Test case when order lines retrieval fails ===");
-    sendEvent(createBody(PO_ID_GET_LINES_INTERNAL_SERVER_ERROR), context.asyncAssertFailure(result -> {
+    sendEvent(createBody(false, PO_ID_GET_LINES_INTERNAL_SERVER_ERROR), context.asyncAssertFailure(result -> {
       assertThat(getPurchaseOrderRetrievals(), hasSize(1));
       assertThat(getPoLineSearches(), nullValue());
       assertThat(getPurchaseOrderUpdates(), nullValue());
@@ -196,7 +194,7 @@ public class CheckInOrderStatusChangeChangeHandlerTest extends ApiTestBase {
   @Test
   public void testOrderUpdateFailure(TestContext context) {
     logger.info("=== Test case when order update fails ===");
-    sendEvent(createBody(PO_ID_OPEN_TO_BE_CLOSED_500_ON_UPDATE), context.asyncAssertFailure(result -> {
+    sendEvent(createBody(false, PO_ID_OPEN_TO_BE_CLOSED_500_ON_UPDATE), context.asyncAssertFailure(result -> {
       assertThat(getPurchaseOrderRetrievals(), hasSize(1));
       assertThat(getPoLineSearches(), hasSize(1));
       assertThat(getPurchaseOrderUpdates(), hasSize(1));
@@ -210,10 +208,10 @@ public class CheckInOrderStatusChangeChangeHandlerTest extends ApiTestBase {
     }));
   }
 
-  private JsonObject createBody(String... ids) {
+  private JsonObject createBody(boolean isItemOrderClose, String... ids) {
     List<JsonObject> orderObjects =
       Stream.of(ids)
-            .map(id -> new JsonObject().put(ORDER_ID, id).put(IS_ITEM_ORDER_CLOSED_PRESENT, false))
+            .map(id -> new JsonObject().put(ORDER_ID, id).put(IS_ITEM_ORDER_CLOSED_PRESENT, isItemOrderClose))
             .collect(Collectors.toList());
     return new JsonObject().put(CheckinHelper.EVENT_PAYLOAD, new JsonArray(orderObjects));
   }
